Fri Nov 19 16:34:16 EST 1999 (Christian)
Fixed a segfault that Paul Radford <prbluebottle@netscape.net>
reported: "Minimize a window into an iconbox, close the icon-
box and minimize the app again". I could not exactly reproduce
this because I already got segfaults when I _entered_ the window
that is uniconified when the iconbox is closed.

In any case, there were strange things going on with
mode.focuswin, because FocusToEWin(NULL) returns immediately
in clickmode ... hope the whole thing is history now.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@1337 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33