Sat Nov 20 19:55:22 GMT 1999 (gilbertt)
Changed the Makefile.am to put epplet images in $EROOT/epplet_data.

I wanted to have epplet subdirs under this directory, but it really
complicated the Makefile, so I left it for now. It was easy for me the make
the changes with sed this way too ;)

Any stuff your app needs to install should go in epplet_data, preferably
with the filename prefixed with the name of your epplet.

Flame me if this bugs you...


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@1345 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33