Wed Nov 24 18:29:19 GMT 1999 (gilbertt)
E-UrlWatch.

Changed size. Added buttons to allow several different "modes" of action.
When you have a url in the X paste buffer, each button will grab the url,
and perform a configurable action upon it. The three are www, ftp, and get.
The default is gnome-moz-remote for the first two, and wget for the third.
The epplet is extremely ugly right now, but I'm working on the code atm, so
ignore that :)

More cool stuff to come :)


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@1389 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33