Tue Dec  7 17:13:06 PST 1999 (KainX)
New dialog for E-NetGraph.  This is the only network epplet I use, so I don't
have any plans to add similar interfaces to the other net epplets, but I
recommend that *someone* does so.

Also, I modified E-Slides and E-NetGraph so that if the config file or the
default settings don't provide a valid image directory (for E-Slides) or a
valid network interface (E-NetGraph), rather than immediately exiting, you
will be presented with the config dialog.  Very handy.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@1544 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33