Polygon filling. Right now only works for convex polygons. Works with a clipping rect, but highly suboptimally (I'm not doing proper polygon clipping here yet, just clipping slowly on each point drawn - really nasty).
There are probably some rounding errors in here. I need to work more on
this, but I have *so* little time for the next few weeks.

Please don't kill me for this code. It's not finished, but I'm about to move
house, and I have to get something working before I pack my PC away.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@3303 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33