After playing many games in 	gint gevas_view_redraw_cb(gpointer data);
I have decided that what young raster was saying about leaving the
"smarts" to evas seems to be the best solution. I tried obscuring
the non visible parts of the evas that the gtkscrolledwindow can not
see and it is actually considerably *slower* with obscures in that case.

Needless to say, the slower version has been disabled, and this is
gevas running as fast as I seem to be able to make it at the moment.

The old code is still in place in gevas_view_redraw_cb() but commented
out in case any rasters or evas gurus want to take a peek at tonites
attempt to supercharge gevas.


There is now a new method to make scrollable gevas'

    // Creates a gevas that is optimized in rendering for a scrolled viewport.
    // Both gevas and scrolledwindow are assigned to new objects with this
    // constructor call.
    void gevas_new_gtkscrolledwindow(GtkgEvas** gevas, GtkWidget** scrolledwindow );

This is to allow me to play games in the future to try to gain render speed for
very large 2^16 square gevas widgets.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@4447 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33