I've changed the way of passing options, because using variadic functions is way too painful when one has to put options together dynamically at runtime (let's say you're passing combinations of 5 options -- you get 2^5 different calls :).
So now I leave it up to the user to put things together manually (efsdsh
does this), or using a variadic convenience wrapper like this:

| v->monitor_id = efsd_start_monitor(e_fs_get_connection(), v->dir,
|				     efsd_ops(2, efsd_op_get_stat(),
|					      efsd_op_get_filetype()));

So simply put the number of options and the option calls in the efsd_ops()
wrapper. Simple enough.

I'll also need to use a hash to keep track of all EfsdFAMMonitors instead
of a list soonish.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@4527 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33