This stuff is not quite up to the level I was hoping to have before committing, but smugg wanted to use some of the new list functions for EWL, so here it goes.
Currently, the list code is pretty good, added the means for adding a free
callback as well as a function for iterating through the list and executing a
specified function for each node's data. Also created single linked lists as
well as the double linked lists that were already in place. The single linked
lists have all the methods as the double except for moving to the previous node.

The AVL tree is also included in this commit and was fairly well tested under
another project I was working on. Unfortunately, I haven't had a chance to test
it thoroughly since modifying it for EWD.

Also, I've included the beginnings of attempts at making the data structures
threadsafe. This is optional and pthreads should not be required to be installed
to use the lib.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@4611 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33