A lot of stuff this time:
* Threads-O-Rama :) -- for now only pthreads are supported. If pthread.h
or libpthread are not found or if you ./configure --disable-threads,
a normal, single-threaded version is built. I think I've been able to
keep the #ifdef clutter at a minimum. There's an access locking
abstraction that allows multiple reads of a data structure but only
single writes (basically a semaphore). The pthread_rwlock_... stuff
is not available in all pthread implementations afaik, so I didn't use
it.

* When compiled with -DDEBUG and -DDEBUG_NEST, you can now turn on and
off the debugging output with the -d and -t options, see efsd --help.

* I had to change some of the stat caching stuff, basically meaning
that getting a stat is now a bit slower than before (because I copy
out struct stats instead of returning a pointer into the hashtable --
I needed this in order to not lock write access to the hashtable longer
than necessary)

* I've therefore tried to reduce the number of stats in efsd_fs.c (the
code for cp, rm etc) a good deal, but that complicates things. I fixed
a couple bugs there, but maybe introduced new ones.

* Added a file magic entry for Blender files.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@4628 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33