I guess it's going by timestamps. Ah well.
After fighting with html on and off for the last week, I've given up on trying
to get the content box to display with edges. I don't remember the tricks
anymore, but, again, I'm totally open to suggestions. :)

Blah. This needs still more work. At least now the ugly background colors
I chose at 5 in the morning are gone. I also like the nav graphics without
the arrows.

If there's problems with this, report the problems, a single time, to the
web team. Thanks.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@4688 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33