fixed filled ellipses - note, this isn't the cleanest fix in the world.
set_cliprect(0,0,0,0) is supposed to disable clipping, but the calls in
api.c subtract one from maxx and maxy, so that breaks some of the logic that
checked for the need to clip. This is a temporary fix, but it does work.

That "whether to clip or not" logic needs a rewrite, but I'm real tired.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@5275 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33