After much hair pulling, fixed a bug in the db loader that affected big endian systems.
When saving compressed images, the header was not being byte swapped, but it
was when loading (which is why db's created on little endian systems worked, but
not ones created and used on the big endian systems). So now the headers are
also swapped to little endian order when saving.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@5458 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33