* Made the lower and raise apply directly so you dont have to    press the apply button.  * Fixed bug causing the images not getting saved.  * Etcher now handles custom states.
This is not tested very much, but im going too boarding school on monday,
so i thought i'd put this in CVS since i wont be able to work much more on it
until the end of next week.

Indentation is changed here also...

If you find that it's acting odd or segv's lemme know and i'll look into it.

Also, since no one want's to be the Etcher maintainer, and i have had the
chance to get familiar with the code, i thought i'd sign up for that post.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@6010 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33