gevas ported to using Evas2 (aka 1.0.0preX) API. This requires GTK+ 2.0.6 by default... monster the configure.in if you want to roll the dice with an earlier GTK+. Some functions in gevas have the 0.6.0 evas API interface, they will probably be left in to attempt to maintain compatibility with old gevas code.
Some methods from Evas2 are still missing, and will be added in time.
You can always use gevasobj_get_evasobj() and gevas_get_evas() to perform
raw Evas API coding. Note that if you do that then call gevas_queue_redraw()
to setup an idle callback to redraw the canvas if you have changed it. Note
also that calling gevas_queue_redraw() 1000 times will only queue the single
actaul redraw for when the GTK main loop gets hit again
(or whenever idle handlers are fired in GTK).

There are still a few buggies but nothing that I have seen will bring down the
house. check out the demo/ dir for a hint as to how to use gevas2. Some of the
examples are trivial tests and testgevas.c is a rather chunky example showing
many things that gevas2 adds to the raw Evas API. When running that demo click
on stuff and drag stuff and rubber band select stuff. (also one of the objects
has a GTK+ popup menu attached to it).

*someone* (which will probably be me) needs to port the enstrom throbbing
icon effect to an evh_ connector, sorry raster that effect is too cool to stay
out of a library ;)

"How soon is now"


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@6369 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33