key grabs are done... ok not as efficiently as i'd liek so having more than a dozen key grabs at any time may impact key event handling a little...
oh yeah.. added to the api .. now theres a modifier mask and a not_mask. the
not mas means "grab the key only if NONE of these modifiers are active and
only if one or more of the mask modifiers are active). using this you can
easily select allmodifiers, none, or a certain set of modifiers. if you need
more than that put in multiple grabs then :) to just have that exact set of
modifiers grabbed have not_mask be the inverse of mask. :)


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@6546 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33