Thu Jun  5 11:15:12 CEST 2003 (Kim)
Unshade if shaded when selecting window on window list. This makes it easier to
retrieve windows lost by accidental shading (and seems to me to be the proper
action anyway).
Attempt #2 to fix fullscreen mode with XFree86-4.3.
Fiddling a bit with _NET_ACTIVE_WINDOW hint handling.
Clean compile of (normally disabled) debug stuff.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@6982 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33