Added etox_get_length(), which returns what should be the strlen of etox_get_text().
Fixed a bug in etox_get_text(), where it was allocating a buffer of et->length + 1, which was too small, since et->length did not include the ending \n's from each line. etox_get_length() was implemented to temporarily fix this.

I think a better solution would be to effectively "unwrap" on get_get(). (We don't actually have to unwrap the lines, just return text without the \n's before soft wrapped lines. Although we need to be careful to increment the buffer pointer by the correct amount.)

Added some strdup()'s to the default marker text / style. Otherwise we'd get malloc errors on free (segv's with efence compiled in).

Use new estyle type field to determine whether bits are obstacles or wrap markers or just normal text. It is no longer assumed that the first non-obstacle bit of a soft-wrapped line is a wrap marker. (Now we can eventually add other wrap marker positions, such as putting a "-" at the end of the line). Also, we should probably add something where wrap markers are positioned in the margins of the text (so the first non wrap marker characters all line up on the left side).


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@7431 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33