use repeat_events, 1; ! mousing in and out of the windows works as expected now.
new group added, it's being used by code on my machine i haven't tested
enough to put in CVS, but it won't hurt the current version by having it
present in the theme, so into cvs it goes!  It's basically for theming user
icons the same way we're currently theming sessions.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@8305 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33