Started a new applications config dialog.  The idea is to drag and drop from the left side to the right side, or to ibar and other places.  Just the skeleton for comment while I make it work.
Some filtering and categorising widgets are needed for the left side.
The right side needs to show the contents of the .order files, and
prevent the user from deleteing standard directories, while enabling
them to create and delete sub menus.  Probably fm is not a good choice for
the right side.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@25172 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33