Add protection code for event objects, avoiding invalid references.
Event objects are meant to be used only inside the callbacks, so any access
after the callback returns will be considered invalid and raise a ValueError.

The first idea (reverted by previous patch) was to make these objects without
any reference to original event (copying data), so they couldn't cause trouble.
But now we will have event_flags field, which is writeable by the callbacks,
so it's better just keep the reference and invalidate the object as we do.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@34119 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33