Introduce imlib_context_disconnect_display().
imlib_context_disconnect_display() should be called when a display
connection is closed but the application continues using imlib2 with a
different display connection.
This is required to avoid to attempt to free cached GCs (in
__imlib_RenderImage) after the associated display connection has been
closed.
It is not unlikely that similar cleanups should be performed elsewhere
in this situation, but the __imlib_RenderImage GCs is the only case I
have found to cause trouble so far.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@35463 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33