Add signature support in eet.
You can use eet -s to sign a eet file and eet -c to check its signature. The
current code doesn't check any certificate trust path, only if the signature
match the certificate and that the certificate could be loaded. If that's not
the case, eet_open will fail. If the file is not signed, eet_open will just
continue to succeed.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@35882 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33