e now supports scaling based on screen dpi (or on a custom scaling factor you can select yourself as a user). also any scaling either calculated from dpi or set by the user is limited to the min/max values a user can impose (thus even if you have very low dpi, e doesn't go below a scaling factor you set - eg 1.0 or 0.5, and same for upscaling).
this should

1. keep all the people who want scalable UI's happy (enable scaling).
2. make e able to (if u want) to adapt to pretty much any dpi screen and
selectively scale just what u want it to.
3. allw you - the user, to be the final controller of just how much to scale
- if you like scaling but think it scales too much - then limit it or set
your own.

right now this will onyl work if your theme supports it... the b&w one does
(mostly). so try that and test. other bits of e need to adapt to this too
(shelf, icon sizes, etc. etc.) as they are not done via edje but code sets
them.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@36211 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33