entrance: better handling of child errors.
If we detect child was killed with a signal or it returned some error
code, notify the syslog. I spent a hour trying to figure out why my
login was not working just to figure out /etc/X11/Xsession didn't
exist, but it was saying everything was fine.

TODO: hint user that command failed.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@36865 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33