* Fix bug with string included in a list found by Viktor Kojouharov. * Add a test case for that one.
Problem was simple type inlining in complex structure (like list/hash/array) is
just a hack. We are creating a subtype with the data chunk. That work with INT
or SHORT for example, but not with STRING because it's a pointer to a STRING and
not directly the STRING. This result in a double pointer dereferencing where it
shouldn't. In fact STRING is not really like other simple data type. So we
should handle it differently.

Still need to fix Array and Hash.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@37024 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33