Couple of fixes for Cython issues and Callback wrapping.
Cython somehow seems to misbehave with "del o[k]" for lists and
dict. It's really weird and I expect to be a bug somewhere else that
was showing there for strange reasons. Since we don't use "del o[k]"
in critical paths, let's replace it with "pop()" version.

Some callbacks were wrapped by using a list of (func, args, kargs) and
the Cython-C function was walking the list and dispatching
there. Problem is that lists could be modified during iteration,
providing weird results. A solution is to copy the list (actually
transform to a tuple, since it's cheaper) and iterate the copy. It's
not that huge performance impact since we often have just few
callbacks for the same signals (see that Evas itself does list walking
and strcmp).



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@37946 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33