e_fm2: view mode is back!
As I suspected it was not error on my code. At the end it was just a
missing theme.

For now I just aliased default to desktop, but I guess it's better to
come with something else transparent so we can see our
wallpaper. Themers?!

PS: for Illume I think it makes sense to have that menu, but for
desktop it does not, maybe we should add more flags to turn off just
view mode selection and keep icon size? Or just turn off lists... or
just let the user choose, there are many strange users out there, one
might like it.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@37983 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33