evas preload: keep the preload thread alive.
before, when no more images were to be preloaded asynchronously, the
thread exited, but were not collected. This leads to a huge leak if
the process is doing aggressive use of image preloading (ie: photo
wall).

collecting dead threads in a proper way (read: without race
conditions) is a bit harder than keeping just one thread alive,
forever. As we do that for evas_pipe (the renderer), let's do the same
with preload and save code.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@38746 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33