fm: handle symbolic links.
Symbolic links are handled similar to "paste". If you "Copy" you'll
get the "Link" menu item that will create a symbolic link instead of
real copying the file.

Drag and Drop works similarly, use "Alt" in order to get a popup and
select "Link" there.

Toma: we do miss icons for it!

Mekius & Ptomaine: please review as I'm not that familiar with e_fm_op.c



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@39170 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33