amey parulekar: patch for notification to concat messages if they have the same id.
Hi,

I've made a couple of changes to the notifications module. Now, if a
notification with a replaces_id equal to the id of an existing notification
is received, then the two are concatenated instead of the original getting
replaced. This is more beneficial for applications like Instant messengers,
where two messages from the same contact will be displayed simultaneously
instead of newer ones replacing older ones.

Thanks,
Amey



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@39933 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33