fix nasty bug of timers not being re-evaluated on rare case.
If there are no other main loop activity than a idlers and one idler
adds a timer, the new (and unique) timer would be ignored since it's
flagged as "just_added" and thus next iteration will not consider it,
possible entering an infinite wait as it could be the only thing to do
in main loop.

Antognolli found this nasty bug while handling timeout-and-die in
Ethumb, where the "disconnect" event is dispatched by EDBus from idler
and it was adding a timer to shutdown the daemon after a while without
clients.

By: Rafael Antognolli <antognolli@profusion.mobi>



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@40923 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33