fix exebuf so it will execute what it should.
Run Command (module exebuf in e) was acting strange for me. I
have .desktops for Firefox and Firefox - Safe mode (well, at least a
menu entries for them), and when I entered "firefox" in the input box,
and choose the command "firefox" from the list under the input field,
exebuf was starting Firefox - Safe Mode (it was the only one showed in
the upper list - "Firefox" was not listed there).
This patch fixes it (the lower selected, upper started problem, not the
"Firefox" is not there one) for me.

By: Tomasz Kontusz


git-svn-id: http://svn.enlightenment.org/svn/e/trunk@40926 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33