start dbus service by its name if it's not up.
we still guess it was up and try to get its name, if not possible then
try to start it and then get is name.

another way to do it would be to always request start and just then
get its name and just then start to listen to NameOwnerChanged. But
the current way is good enough and should save some roundtrips for
good cases (server is up).



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@40955 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33