start service by name requires one more parameter (flags).
The dbus api has a "flags" parameter that is now unused but may be in
future, it was missing and dbus was giving method mismatch.

I forgot to commit this but changed Ethumb_Client, then Viktor "fixed"
it by reverting such change. Now going back to my code and adding "0"
as flag.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@40968 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33