Illume now uses thresholds to avoid unwanted clicks.
Use the e_fm2 single_click_delay option in illume's e_fm2 to allow to
 configure, from the launcher settings, a click delay to prevent
un-wanted clicks.

I've added a slider to configure the delay from the settings (also if it
has some issues, check the code comments and let me know if should be
redesigned). The default value is set to 120 ms, but it works great too
for lower values.

By: Marco Trevisan (Treviño)



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@41128 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33