Goodbye stupid setuptool problems with Cython.
setuptools is maybe one of the most stupid tools out there, it patches
distutils aiming to "fix" it, but it actually make things worse. One
example is that although both Pyrex and Cython (an old Pyrex fork that
is now much more active and maintained), it will check just for Pyrex
and break distutils, that works perfectly with Cython!

My trick is to force Cython module into sys.modules as the name
'Pyrex', before setuptools is imported and patches stuff. This works
nicely, but is ugly as hell.

So far setuptools usefulness is quite limited. I'm thinking about
going straight to distutils or even automake/autoconf, so we can avoid
those hacks around pkgconfig.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@46149 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33